cleanup
split rbac tests

+ api improvements

replace by beforeAll by RhdhApi

delete redundant toBeTruthy

moving constants at top of methods

fix RhdhApi api url
